+++ plugins/admission/pkg/models/models.go
@@ -1,0 +1,38 @@
+package models
+
+import (
+	"github.com/fairwindsops/insights-plugins/opa/pkg/opa"
+	polarisconfiguration "github.com/fairwindsops/polaris/pkg/config"
+)
+
+// ScoreOutOfBoundsMessage is the message for the error when the score returned by Insights is out of bounds.
+const ScoreOutOfBoundsMessage = "score out of bounds"
+
+// Resource represents a Kubernetes resource with information about what file it came from.
+type Resource struct {
+	Kind      string
+	Name      string
+	Filename  string
+	Namespace string
+}
+
+// ReportInfo is the information about a run of one of the reports.
+type ReportInfo struct {
+	Report   string
+	Version  string
+	Contents []byte
+}
+
+// Configuration saves any config from Insights.
+type Configuration struct {
+	Reports struct {
+		Polaris bool
+		Pluto   bool
+		OPA     bool
+	}
+	OPA struct {
+		CustomChecks         []opa.OPACustomCheck
+		CustomCheckInstances []opa.CheckSetting
+	}
+	Polaris *polarisconfiguration.Configuration
+}