+++ plugins/uploader/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.2
+
+* Decrease verbosity of logging unless DEBUG environment variable is set.
+* Fix bug
+* Increase verbosity of timeout message
+
 ## 0.2.0
 
 * Send error logs when a container fails.